Move our custom logger to logging.py
Also remove bioimg's own custom logger.
