+++ src/tiledb/cloud/bioimg/helpers.py
@@ -1,31 +1,32 @@
-import logging
+# import logging
 from typing import Sequence
 
 import tiledb
-from tiledb.cloud.utilities import get_logger
+
+# from tiledb.cloud.logging import get_logger
 
 
-def get_logger_wrapper(
-    verbose: bool = False,
-) -> logging.Logger:
-    """
-    Get a logger instance and log version information.
+# def get_logger_wrapper(
+#     verbose: bool = False,
+# ) -> logging.Logger:
+#     """
+#     Get a logger instance and log version information.
 
-    :param verbose: verbose logging, defaults to False
-    :return: logger instance
-    """
+#     :param verbose: verbose logging, defaults to False
+#     :return: logger instance
+#     """
 
-    level = logging.DEBUG if verbose else logging.INFO
-    logger = get_logger(level)
+#     level = logging.DEBUG if verbose else logging.INFO
+#     logger = get_logger(level)
 
-    logger.debug(
-        "tiledb.cloud=%s, tiledb=%s, libtiledb=%s",
-        tiledb.cloud.version.version,
-        tiledb.version(),
-        tiledb.libtiledb.version(),
-    )
+#     logger.debug(
+#         "tiledb.cloud=%s, tiledb=%s, libtiledb=%s",
+#         tiledb.cloud.version.version,
+#         tiledb.version(),
+#         tiledb.libtiledb.version(),
+#     )
 
-    return logger
+#     return logger
 
 
 def serialize_filter(filter):
