+++ src/tiledb/cloud/geospatial/ingestion.py
@@ -20,12 +20,12 @@ import attrs
 
 import tiledb
 from tiledb.cloud.files import utils as file_utils
+from tiledb.cloud.logging import get_logger_wrapper
 from tiledb.cloud.utilities import Profiler
 from tiledb.cloud.utilities import as_batch
 from tiledb.cloud.utilities import chunk
 from tiledb.cloud.utilities import create_log_array
 from tiledb.cloud.utilities import find
-from tiledb.cloud.utilities import get_logger_wrapper
 from tiledb.cloud.utilities import max_memory_usage
 from tiledb.cloud.utilities import run_dag
