+++ src/tiledb/cloud/utilities/__init__.py
@@ -1,8 +1,6 @@
 from ._common import as_batch
 from ._common import chunk
 from ._common import find
-from ._common import get_logger
-from ._common import get_logger_wrapper
 from ._common import max_memory_usage
 from ._common import process_stream
 from ._common import read_aws_config
@@ -13,6 +11,8 @@ from ._common import set_aws_context
 from .consolidate import consolidate_and_vacuum
 from .consolidate import consolidate_fragments
 from .consolidate import group_fragments
+from .logging import get_logger
+from .logging import get_logger_wrapper
 from .profiler import Profiler
 from .profiler import create_log_array
 from .profiler import write_log_event
